orchestration/tools.py
"""
Tools and utilities for Databricks workspace integration.
Handles connections to serving endpoints, SQL queries, and metadata retrieval.
"""

import os
from typing import Optional, Dict, Any, List
from dotenv import load_dotenv

load_dotenv()


class DatabricksConfig:
    """Configuration for Databricks workspace connection"""

    def __init__(self):
        self.workspace_url = os.getenv("DATABRICKS_HOST", "")
        self.token = os.getenv("DATABRICKS_TOKEN", "")
        self.workspace_id = os.getenv("DATABRICKS_WORKSPACE_ID", "")

    def is_configured(self) -> bool:
        """Check if Databricks is properly configured"""
        return bool(self.workspace_url and self.token)


class DatabricksEndpointTool:
    """
    Tool for calling Databricks Model Serving endpoints.
    Maps to your existing workspace/endpoint configuration.
    """

    def __init__(self):
        self.config = DatabricksConfig()
        self.endpoints = {}  # Populated from agents-db/data/store.js

    def register_endpoint(self, name: str, endpoint_id: str, workspace_url: str):
        """Register a Databricks serving endpoint"""
        self.endpoints[name] = {
            "endpoint_id": endpoint_id,
            "workspace_url": workspace_url,
            "status": "pending"
        }

    async def call_endpoint(
        self,
        endpoint_name: str,
        query: str,
        parameters: Optional[Dict[str, Any]] = None
    ) -> Dict[str, Any]:
        """
        Call a registered Databricks endpoint.

        Args:
            endpoint_name: Name of registered endpoint
            query: Input query or prompt
            parameters: Optional parameters for the endpoint

        Returns:
            Response from the endpoint
        """
        if endpoint_name not in self.endpoints:
            return {
                "error": f"Endpoint '{endpoint_name}' not found",
                "available_endpoints": list(self.endpoints.keys())
            }

        endpoint_config = self.endpoints[endpoint_name]

        # TODO: Implement HTTP call to Databricks endpoint
        # This will be replaced with actual databricks-langchain integration

        response = {
            "endpoint": endpoint_name,
            "status": "success",
            "response": f"Response from {endpoint_name}",
            "latency_ms": 0
        }

        return response


class DatabricksWorkspaceTool:
    """
    Tool for querying Databricks workspace metadata.
    Lists tables, endpoints, clusters, models, etc.
    """

    def __init__(self):
        self.config = DatabricksConfig()

    async def list_tables(self, catalog: str = "main", schema: str = "default") -> List[str]:
        """List Delta tables in a catalog/schema"""
        # TODO: Implement using Databricks SQL API
        return []

    async def list_endpoints(self) -> List[Dict[str, Any]]:
        """List all Model Serving endpoints in workspace"""
        # TODO: Implement using Databricks API
        return []

    async def get_endpoint_status(self, endpoint_id: str) -> Dict[str, Any]:
        """Get status of a specific endpoint"""
        # TODO: Implement using Databricks API
        return {
            "endpoint_id": endpoint_id,
            "status": "READY",
            "creation_timestamp": 0
        }

    async def query_workspace(self, sql: str) -> List[Dict[str, Any]]:
        """Execute a SQL query on Databricks workspace"""
        # TODO: Implement using Databricks SQL warehouse
        return []


class RAGTool:
    """
    Tool for Retrieval-Augmented Generation.
    Searches documentation and past solutions.
    """

    def __init__(self):
        self.vectors = []  # Will be populated from Vector Search
        self.config = DatabricksConfig()

    async def search_documentation(self, query: str, top_k: int = 5) -> List[Dict[str, Any]]:
        """Search documentation for relevant information"""
        # TODO: Implement using Databricks Vector Search
        # This requires a Vector Search index on your documentation
        return [
            {
                "source": "documentation",
                "content": "Sample documentation snippet",
                "relevance_score": 0.85,
                "document_id": "doc_1"
            }
        ]

    async def search_past_solutions(self, query: str, top_k: int = 5) -> List[Dict[str, Any]]:
        """Search for similar past solutions"""
        # TODO: Implement using MLflow Model Registry or Delta tables
        return [
            {
                "source": "past_solution",
                "solution_id": "sol_1",
                "similarity_score": 0.78,
                "description": "Similar solution from past project"
            }
        ]


# Global tool instances
endpoint_tool = DatabricksEndpointTool()
workspace_tool = DatabricksWorkspaceTool()
rag_tool = RAGTool()


def initialize_tools():
    """Initialize all tools with configuration from agents-db"""
    # TODO: Load endpoint configuration from /data/store.js
    # Example:
    # endpoint_tool.register_endpoint(
    #     "solution_designer",
    #     "solution-designer-endpoint",
    #     "https://your-workspace.databricks.com"
    # )
    pass
