orchestration/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
langgraph>=0.0.65
langchain-core>=0.2.0
langchain>=0.1.7
langchain-google-genai>=0.0.1
google-generativeai>=0.3.0
mlflow>=2.9.0
python-dotenv>=1.0.0
httpx>=0.25.0
aiohttp>=3.9.0
tenacity>=8.2.0
